DS-18 add get location list feature

+++ backend/DirectoryService/src/DirectoryService.Shared/Results/PagedResult.cs
@@ -4,5 +4,5 @@ public sealed record PagedResult<T>(
     IEnumerable<T> Results,
     int Page, 
     int PageSize, 
-    int TotalCount
+    long TotalCount
 );
